+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to
 - ✨(user) allow to use conversation data for analytics #23
 - ✨(chat) enforce response in user language #24
 - 📈(langfuse) add light instrumentation #26
+- 🚑️(agent) allow Mistral w/ vLLM & tools #36
 
 
 [unreleased]: https://github.com/numerique-gouv/conversations/compare/HEAD...main